requirements.txt
accelerate==0.30.1
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
blinker==1.4
cbor==1.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
einops==0.8.1
faiss-gpu-cu12==1.8.0.2
fastapi==0.119.0
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httplib2==0.20.2
httptools==0.7.1
huggingface-hub==0.35.3
idna==3.11
ijson==3.4.0.post0
importlib-metadata==4.6.4
inscriptis==2.6.0
ir_datasets==0.5.11
ir_measures==0.4.1
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==6.0.2
lz4==4.4.4
MarkupSafe==3.0.3
more-itertools==8.10.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.12.1
pydantic_core==2.41.3
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytrec_eval-terrier==0.5.9
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
SecretStorage==3.3.1
sentence-transformers==5.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.0
trec-car-tools==2.6
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
wadllib==1.3.6
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.0
websockets==15.0.1
zipp==1.0.0
zlib-state==0.1.10
zstandard==0.25.0
